Q2 Description
Question - Simulate Cyclic Redundancy Check (CRC) error detection algorithm for noisy channel. 

CRC or Cyclic Redundancy Check is a method of detecting accidental changes/errors in the communication channel. CRC uses Generator Polynomial which is available on both sender and receiver side. An example generator polynomial is of the form like x3 + x + 1. This generator polynomial represents key 1011. 

Both ends perform modulo-2 division in the end when in last receiver end perform modulo-2 division and if the remainder is 0, then there are no errors.  However the remainder is not all zeroes, the error is detected at the receiver side. 

This is proved in above program where we perform Simulate Cyclic Redundancy Check (CRC) error detection algorithm for noisy channel.