Q1 Description
Q - Simulate and implement go back n sliding window protocol. 

Go back and N is an algorithm that is a form of the sliding window protocol. In this algorithm, we have a fixed size window which is sending the frames in a window. Whenever the frames are lost in a Window the entire window is sent again. and It does not matter whether the previous frames were received or not.

Both sender and receiver agrees on some window size. If window size=w then after sending w frames sender waits for the acknowledgement (ack) of the first frame. In sliding window protocol the receiver has to have some memory to compensate any loss in transmission or if the frames are received unordered. 

In above Program we can see clearly how go back n sliding window protocol works.